Move code creating links to a separate file
